+++ dart_lmdb2/CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 0.9.8
 
 * iOS: Added support for iOS simulators with architecture-specific builds
-* iOS: Created universal simulator library with both x86_64 and arm64 architectures
-* iOS: Reorganized libraries into separate device and simulator directories
+*      Created universal simulator library with both x86_64 and arm64 architectures
+*      Reorganized libraries into separate device and simulator directories
+* Android: Fixed dynamic library loading to use proper package path resolution
+* Release: Download binaries at rumtime from GitHub from correpsonding release page
 
 ## 0.9.7